requirements.txt
yt-dlp>=2023.3.4
requests>=2.28.2
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6
2captcha-python>=1.1.3
PySimpleGUI>=4.60.4  # Optional for GUI